diplom/requirements.txt
selenium==4.15.0
requests==2.31.0
pytest==8.4.2
allure-pytest==2.15.0
webdriver-manager==4.0.1